Fix and improve the betting dashboard display and data handling
Update dashboard layout and data loading logic, fix display of league names in match summary, and ensure proper handling of empty dataframes for PnL distribution. Also, convert numerical data to floats for consistency.

+++ data_loader.py
@@ -209,8 +209,8 @@ class DataLoader:
             max_risk_per_match = 0.0
         
         return {
-            'bankroll': bankroll,
-            'total_risk': total_risk,
-            'risk_percentage': (total_risk / bankroll * 100) if bankroll > 0 else 0,
-            'max_risk_per_match': max_risk_per_match
+            'bankroll': float(bankroll),
+            'total_risk': float(total_risk),
+            'risk_percentage': float((total_risk / bankroll * 100) if bankroll > 0 else 0),
+            'max_risk_per_match': float(max_risk_per_match)
         }
